+++ CustomRunOptions/ExternalAPIs/IMoreDifficultiesApi.cs
@@ -1,0 +1,6 @@
+﻿namespace TheJazMaster.MoreDifficulties;
+
+public interface IMoreDifficultiesApi
+{
+	public StarterDeck? GetAltStarters(Deck deck);
+}